修改告警

+++ frameworks/base/ressched/ressched_report.cpp
@@ -255,7 +255,7 @@ void ResSchedReport::HandleTouchPullMove(const TouchEvent& touchEvent)
     RecordTouchEvent(touchEvent);
 }
 
-float ResSchedReport::GetUpVelocity(const TouchEvent& lastMoveInfo,
+double ResSchedReport::GetUpVelocity(const TouchEvent& lastMoveInfo,
     const TouchEvent& upEventInfo)
 {
     float distance = sqrt(pow(lastMoveInfo.x - upEventInfo.x, SQUARE) + pow(lastMoveInfo.y - upEventInfo.y, SQUARE));
@@ -264,7 +264,7 @@ float ResSchedReport::GetUpVelocity(const TouchEvent& lastMoveInfo,
     if (time < MILL_SECOND_UNIT) {
         return 0.0f;
     }
-    return distance * dpi_ / (time / MILL_SECOND_UNIT); //unit: pixel/ms
+    return (double)distance * dpi_ / (double)(time / MILL_SECOND_UNIT); //unit: pixel/ms
 }
 
 void ResSchedReport::LoadPageEvent(int32_t value)
